Make the backend very permissive wrt CORS
To enable local development, as the API is served from a different host
(well, port) than the web frontend is, we need to set some CORS headers
and respond appropriately to CORS preflight requests. This is something
we're already doing in production via the use of NGINX, so this is a
non-change there.